+++ crates/bls-gadgets/src/bitmap.rs
@@ -1,4 +1,4 @@
-use crate::{is_setup, smaller_than::SmallerThanGadget};
+use crate::{utils::is_setup, smaller_than::SmallerThanGadget};
 use algebra::PrimeField;
 use r1cs_core::{ConstraintSystem, LinearCombination, SynthesisError};
 use r1cs_std::{
