+++ crates/bls-gadgets/src/hash_to_group.rs
@@ -1,4 +1,7 @@
-use crate::{bits_to_bytes, bytes_to_bits, constrain_bool, is_setup, YToBitGadget};
+use crate::{
+    utils::{bits_to_bytes, bytes_to_bits, constrain_bool, is_setup},
+    YToBitGadget,
+};
 use bls_crypto::{
     hashers::{
         composite::{CompositeHasher, CRH},
